Marid Code (#35)
* Initial push

This is a list of all of the code that I've done so far for the game. Please ignore the chargen.py file. It's a bit broken as I'm trying to implement the rest of the data files into it.

* Reality Levels & More

Several updates in this one:
roll/log on rooms
language now has commands to +language/add
There are three fae-specific commands:
+flook (look at a fae desc)
+fpose (pose something that only fae can see)
+femit (emit something that only fae can see)
people in the Umbra and Material realm should not see each others' poses, emits, says, or OOC communication.
People in the Umbra or Material realm can also not use 'look' to see each other. It will return a 'Could not find' message.
Characters can set themselves as unfindable or remove it (+unfindable on/off)
+where accounts for unfindable characters by listing them in an 'unfindable characters' section. It will also indicate if someone's in the Umbra.
Additionally, includes weather commands that shows current 1:1 weather in San Diego, including date, time, high tide, low tide, humidity, and forecast for tomorrow.

* Jobs and more

This includes the modified +jobs combination system, the basics for equipment, some additional data pool stuff, the basics for mortal+ stuff and some bugfixes.