Merge dev to main (#39)
* Adding  CSRF Trusted origins

* Adding servername info

* fixed upload  script.

* Added pose commands, update room parent, basic stats commands.  MOAR settings.

* Updated stats, stat setting

* Updates to room parent

* Bbs into Dev (#8)

* Initial Commit, Not All Cmds Fully Tested

* Re-organized, Added Unit Tests

* Added ReadMe.MD

* Fixed merge mistake (Cache caused me to piss it)

* Updates to Stats Sheet an Damage

* Script typeclass fix

* bbs fix

* Fixes for Sheet/Stats/locks

* Gradientname updates

* updates to some command permissions.

* Gradient name updates

* Dice Rolls

* Dice Roller + Language

* Fix for exit aliases.

* Language Exits fix

* Bug fix, formatting.

* +Language code, Exit Fix, Cjarater.py addotopms/edits

* Emit fix.

* fixing pose b ug

* pose bug fix

* bug fix

* bug fix

* bug fix

* Added at_create  on Chracter ogject. (Thanks Tell!!!!!)

* Fixing get_languages

* Chargen code [for testing]

* Initial Commit, District, Sector, Neighborhood, Site

* Modified to use typeclasses and excluding crisis for now

* working on +shift

* Added color settings

* Updates & bug fixes.  Added +where.

* settings update

* update settings

* updating settings.py

* updating settings

* settings update

* Initial push

This is a list of all of the code that I've done so far for the game. Please ignore the chargen.py file. It's a bit broken as I'm trying to implement the rest of the data files into it.

* Reality Levels & More

Several updates in this one:
roll/log on rooms
language now has commands to +language/add
There are three fae-specific commands:
+flook (look at a fae desc)
+fpose (pose something that only fae can see)
+femit (emit something that only fae can see)
people in the Umbra and Material realm should not see each others' poses, emits, says, or OOC communication.
People in the Umbra or Material realm can also not use 'look' to see each other. It will return a 'Could not find' message.
Characters can set themselves as unfindable or remove it (+unfindable on/off)
+where accounts for unfindable characters by listing them in an 'unfindable characters' section. It will also indicate if someone's in the Umbra.
Additionally, includes weather commands that shows current 1:1 weather in San Diego, including date, time, high tide, low tide, humidity, and forecast for tomorrow.

* Jobs and more

This includes the modified +jobs combination system, the basics for equipment, some additional data pool stuff, the basics for mortal+ stuff and some bugfixes.

* Updates and Fixes to Marid Merge

* Update to settings

* Bug Fix models.py

* Update RoomParent

* update RoomParent

* Updating requirements, finishing install build

* adding migrations back.

* Update Rooms

This fixes the roll log that's saved in the room, and also will adjust the tabs and carriage returns to allow for either lower or uppercase r/R, t/T (similar to MUSH or MUX)

* Starting wiki project.  Refreshing web static files.

* continued work on wiki

* comtinued work on wiki

* fixing migrations

* updates to wiki

* fix settings

* settings update

* config changes

* updates for featured articles.

* css updates

* updates to wiki context menus

* Marid (#38)

* Room bug fix

Fixes bug where players can't see themselves or others in the room.

* Umbra/Room fixes

This fixes the in_umbra state for characters, adjusts the room coloration based on what realm you're in, and removes the doubled message about stepping into or out of the Umbra.

* Updated OOC command

Updated OOC command to default to the room 'Central District' when using it without having used +ic.

---------

Co-authored-by: Lem Canady <lem.canady@gmail.com>
Co-authored-by: vectortangent <kent.gillenwater@gmail.com>
Co-authored-by: Daniel K <daniel.kik.adv@protonmail.com>
Co-authored-by: Admin <admin@127.0.0.1>